refactor: move analyticstools hooks to shared homeview and generalize (#1598)
