--- app/components/Home/components/Section/components/SectionAnalytics/components/AnalyticsTools/hooks/useInteractiveAnalytics/common/contants.ts
@@ -1,1 +1,0 @@
-export const ROWS = 1;
